requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
cryptography==44.0.2
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.4.1
django-cors-headers==4.7.0
django-extensions==3.2.3
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docker==7.1.0
drf-yasg==1.21.9
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jiter==0.8.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
openai==1.65.4
packaging==24.2
pillow==11.1.0
psycopg==3.2.5
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.44.0
whitenoise==6.9.0
